+++ package/src/shared/interfaces/AudioPlayerOptions.ts
@@ -1,0 +1,6 @@
+export interface AudioPlayerOptions {
+  loop: boolean;
+  volume: number;
+  // isSpatial: boolean;
+  // position: Vector2;
+}