+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2024-11-20
+
+### Added
+
+- AudioManager: Back & front, to store audio data.
+- AudioPlayer: Front only, to play/pause/resume/stop/... audio.
+- MemView Back: Add audio methods.
+- MemView Front: Add audio callbacks.
+
 ## [0.4.4] - 2024-11-14
 
 ### Added (minor changes)